Add not working tests

+++ src/main/java/com/ihorchyzh/app/classwork/lesson1/WideningCasting.java
@@ -5,7 +5,7 @@ package com.ihorchyzh.app.classwork.lesson1;
  */
 public class WideningCasting {
 
-    public double widening(int intNumber){
+    public float widening(int intNumber){
 
         float floatNumber = intNumber;
         return floatNumber;
